// src/api/model/getWorkshopParams.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * arsa-backend-rs
 * OpenAPI spec version: 0.1.0
 */
import type { GetWorkshopSort } from './getWorkshopSort';

export type GetWorkshopParams = {
    /**
     * Free-text search string
     * @nullable
     */
    search?: string | null;
    /**
 * One or more tags to filter by (repeat the parameter for multiple tags,
e.g. `?tags=VEHICLES&tags=COMPOSITIONS`)
 */
    tags?: string[];
    sort?: GetWorkshopSort;
    /**
     * Page number (1-based, default: 1)
     * @minimum 0
     */
    page?: number;
};