// src/api/model/getWorkshopSort.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * arsa-backend
 * OpenAPI spec version: 2.0.0
 */

export type GetWorkshopSort = (typeof GetWorkshopSort)[keyof typeof GetWorkshopSort];

export const GetWorkshopSort = {
    newest: 'newest',
    popularity: 'popularity',
    subscribers: 'subscribers',
    version_size: 'version_size'
} as const;